mpi/datatype: split MPIR implementations from MPI functions
