+++ src/mpi/datatype/type_contiguous.c
@@ -25,164 +25,6 @@ int MPI_Type_contiguous(int count, MPI_Datatype oldtype, MPI_Datatype * newtype)
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Type_contiguous
 #define MPI_Type_contiguous PMPI_Type_contiguous
-
-int MPIR_Type_contiguous(int count, MPI_Datatype oldtype, MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Datatype *new_dtp;
-
-    if (count == 0)
-        return MPII_Type_zerolen(newtype);
-
-    /* allocate new datatype object and handle */
-    new_dtp = (MPIR_Datatype *) MPIR_Handle_obj_alloc(&MPIR_Datatype_mem);
-    /* --BEGIN ERROR HANDLING-- */
-    if (!new_dtp) {
-        mpi_errno = MPIR_Err_create_code(MPI_SUCCESS, MPIR_ERR_RECOVERABLE,
-                                         "MPIR_Type_contiguous",
-                                         __LINE__, MPI_ERR_OTHER, "**nomem", 0);
-        return mpi_errno;
-    }
-    /* --END ERROR HANDLING-- */
-
-    /* handle is filled in by MPIR_Handle_obj_alloc() */
-    MPIR_Object_set_ref(new_dtp, 1);
-    new_dtp->is_committed = 0;
-    new_dtp->attributes = NULL;
-    new_dtp->name[0] = 0;
-    new_dtp->contents = NULL;
-    new_dtp->flattened = NULL;
-
-    new_dtp->typerep.handle = NULL;
-
-    if (HANDLE_IS_BUILTIN(oldtype)) {
-        MPI_Aint el_sz = MPIR_Datatype_get_basic_size(oldtype);
-
-        new_dtp->size = count * el_sz;
-        new_dtp->true_lb = 0;
-        new_dtp->lb = 0;
-        new_dtp->true_ub = count * el_sz;
-        new_dtp->ub = new_dtp->true_ub;
-        new_dtp->extent = new_dtp->ub - new_dtp->lb;
-
-        new_dtp->alignsize = el_sz;
-        new_dtp->n_builtin_elements = count;
-        new_dtp->builtin_element_size = el_sz;
-        new_dtp->basic_type = oldtype;
-        new_dtp->is_contig = 1;
-    } else {
-        /* user-defined base type (oldtype) */
-        MPIR_Datatype *old_dtp;
-
-        MPIR_Datatype_get_ptr(oldtype, old_dtp);
-
-        new_dtp->size = count * old_dtp->size;
-
-        MPII_DATATYPE_CONTIG_LB_UB((MPI_Aint) count,
-                                   old_dtp->lb,
-                                   old_dtp->ub, old_dtp->extent, new_dtp->lb, new_dtp->ub);
-
-        /* easiest to calc true lb/ub relative to lb/ub; doesn't matter
-         * if there are sticky lb/ubs or not when doing this.
-         */
-        new_dtp->true_lb = new_dtp->lb + (old_dtp->true_lb - old_dtp->lb);
-        new_dtp->true_ub = new_dtp->ub + (old_dtp->true_ub - old_dtp->ub);
-        new_dtp->extent = new_dtp->ub - new_dtp->lb;
-
-        new_dtp->alignsize = old_dtp->alignsize;
-        new_dtp->n_builtin_elements = count * old_dtp->n_builtin_elements;
-        new_dtp->builtin_element_size = old_dtp->builtin_element_size;
-        new_dtp->basic_type = old_dtp->basic_type;
-
-        MPIR_Datatype_is_contig(oldtype, &new_dtp->is_contig);
-    }
-
-    mpi_errno = MPIR_Typerep_create_contig(count, oldtype, new_dtp);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    *newtype = new_dtp->handle;
-
-    MPL_DBG_MSG_P(MPIR_DBG_DATATYPE, VERBOSE, "contig type %x created.", new_dtp->handle);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Type_contiguous_impl(int count, MPI_Datatype oldtype, MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Datatype *new_dtp;
-    MPI_Datatype new_handle;
-
-    mpi_errno = MPIR_Type_contiguous(count, oldtype, &new_handle);
-
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-
-    MPIR_Datatype_get_ptr(new_handle, new_dtp);
-    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_CONTIGUOUS, 1, /* ints (count) */
-                                           0, 1, &count, NULL, &oldtype);
-
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-
-    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-
-    goto fn_exit;
-}
-
-int MPIR_Type_contiguous_x_impl(MPI_Count count, MPI_Datatype oldtype, MPI_Datatype * newtype)
-{
-    /* to make 'count' fit MPI-3 type processing routines (which take integer
-     * counts), we construct a type consisting of N INT_MAX chunks followed by
-     * a remainder.  e.g for a count of 4000000000 bytes you would end up with
-     * one 2147483647-byte chunk followed immediately by a 1852516353-byte
-     * chunk */
-    MPI_Datatype chunks, remainder;
-    MPI_Aint lb, extent, disps[2];
-    int blocklens[2];
-    MPI_Datatype types[2];
-    int mpi_errno;
-
-    /* truly stupendously large counts will overflow an integer with this math,
-     * but that is a problem for a few decades from now.  Sorry, few decades
-     * from now! */
-    MPIR_Assert(count / INT_MAX == (int) (count / INT_MAX));
-    int c = (int) (count / INT_MAX);    /* OK to cast until 'count' is 256 bits */
-    int r = count % INT_MAX;
-
-    mpi_errno = MPIR_Type_vector_impl(c, INT_MAX, INT_MAX, oldtype, &chunks);
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-    mpi_errno = MPIR_Type_contiguous_impl(r, oldtype, &remainder);
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-
-    MPIR_Type_get_extent_impl(oldtype, &lb, &extent);
-
-    blocklens[0] = 1;
-    blocklens[1] = 1;
-    disps[0] = 0;
-    disps[1] = c * extent * INT_MAX;
-    types[0] = chunks;
-    types[1] = remainder;
-
-    mpi_errno = MPIR_Type_create_struct_impl(2, blocklens, disps, types, newtype);
-
-    MPIR_Type_free_impl(&chunks);
-    MPIR_Type_free_impl(&remainder);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
 #endif
