+++ src/mpi/datatype/type_create_indexed_block.c
@@ -24,48 +24,6 @@ int MPI_Type_create_indexed_block(int count, int blocklength, const int array_of
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Type_create_indexed_block
 #define MPI_Type_create_indexed_block PMPI_Type_create_indexed_block
-
-int MPIR_Type_create_indexed_block_impl(int count,
-                                        int blocklength,
-                                        const int array_of_displacements[],
-                                        MPI_Datatype oldtype, MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_CHKLMEM_DECL(1);
-    MPI_Datatype new_handle;
-    MPIR_Datatype *new_dtp;
-    int i, *ints;
-
-    mpi_errno = MPIR_Type_blockindexed(count, blocklength, array_of_displacements, 0,   /* dispinbytes */
-                                       oldtype, &new_handle);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIR_CHKLMEM_MALLOC_ORJUMP(ints, int *, (count + 2) * sizeof(int), mpi_errno,
-                               "content description", MPL_MEM_BUFFER);
-
-    ints[0] = count;
-    ints[1] = blocklength;
-
-    for (i = 0; i < count; i++)
-        ints[i + 2] = array_of_displacements[i];
-
-    MPIR_Datatype_get_ptr(new_handle, new_dtp);
-    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_INDEXED_BLOCK, count + 2,      /* ints */
-                                           0,   /* aints */
-                                           1,   /* types */
-                                           ints, NULL, &oldtype);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
-
-  fn_exit:
-    MPIR_CHKLMEM_FREEALL();
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-
 #endif
 
 /*@
