+++ src/mpi/datatype/type_create_resized.c
@@ -27,92 +27,6 @@ int MPI_Type_create_resized(MPI_Datatype oldtype, MPI_Aint lb, MPI_Aint extent,
 
 #endif
 
-/* --BEGIN ERROR HANDLING-- */
-static int MPII_Type_create_resized_memory_error(void)
-{
-    int mpi_errno;
-
-    mpi_errno = MPIR_Err_create_code(MPI_SUCCESS,
-                                     MPIR_ERR_RECOVERABLE,
-                                     "MPIR_Type_create_resized",
-                                     __LINE__, MPI_ERR_OTHER, "**nomem", 0);
-    return mpi_errno;
-}
-
-/* --END ERROR HANDLING-- */
-
-int MPIR_Type_create_resized(MPI_Datatype oldtype,
-                             MPI_Aint lb, MPI_Aint extent, MPI_Datatype * newtype_p)
-{
-    MPIR_Datatype *new_dtp;
-
-    new_dtp = (MPIR_Datatype *) MPIR_Handle_obj_alloc(&MPIR_Datatype_mem);
-    /* --BEGIN ERROR HANDLING-- */
-    if (!new_dtp)
-        return MPII_Type_create_resized_memory_error();
-    /* --END ERROR HANDLING-- */
-
-    /* handle is filled in by MPIR_Handle_obj_alloc() */
-    MPIR_Object_set_ref(new_dtp, 1);
-    new_dtp->is_committed = 0;
-    new_dtp->attributes = 0;
-    new_dtp->name[0] = 0;
-    new_dtp->contents = 0;
-    new_dtp->flattened = NULL;
-
-    new_dtp->typerep.handle = NULL;
-
-    /* if oldtype is a basic, we build a contiguous typerep of count = 1 */
-    if (HANDLE_IS_BUILTIN(oldtype)) {
-        int oldsize = MPIR_Datatype_get_basic_size(oldtype);
-
-        new_dtp->size = oldsize;
-        new_dtp->true_lb = 0;
-        new_dtp->lb = lb;
-        new_dtp->true_ub = oldsize;
-        new_dtp->ub = lb + extent;
-        new_dtp->extent = extent;
-        new_dtp->alignsize = oldsize;   /* FIXME ??? */
-        new_dtp->n_builtin_elements = 1;
-        new_dtp->builtin_element_size = oldsize;
-        new_dtp->is_contig = (extent == oldsize) ? 1 : 0;
-        new_dtp->basic_type = oldtype;
-    } else {
-        /* user-defined base type */
-        MPIR_Datatype *old_dtp;
-
-        MPIR_Datatype_get_ptr(oldtype, old_dtp);
-
-        new_dtp->size = old_dtp->size;
-        new_dtp->true_lb = old_dtp->true_lb;
-        new_dtp->lb = lb;
-        new_dtp->true_ub = old_dtp->true_ub;
-        new_dtp->ub = lb + extent;
-        new_dtp->extent = extent;
-        new_dtp->alignsize = old_dtp->alignsize;
-        new_dtp->n_builtin_elements = old_dtp->n_builtin_elements;
-        new_dtp->builtin_element_size = old_dtp->builtin_element_size;
-        new_dtp->basic_type = old_dtp->basic_type;
-
-        if (extent == old_dtp->size)
-            MPIR_Datatype_is_contig(oldtype, &new_dtp->is_contig);
-        else
-            new_dtp->is_contig = 0;
-    }
-
-    int mpi_errno = MPIR_Typerep_create_resized(oldtype, lb, extent, new_dtp);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    *newtype_p = new_dtp->handle;
-
-    MPL_DBG_MSG_P(MPIR_DBG_DATATYPE, VERBOSE, "resized type %x created.", new_dtp->handle);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 /*@
    MPI_Type_create_resized - Create a datatype with a new lower bound and
      extent from an existing datatype
