+++ src/mpi/datatype/type_create_struct.c
@@ -25,48 +25,6 @@ int MPI_Type_create_struct(int count, const int array_of_blocklengths[],
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Type_create_struct
 #define MPI_Type_create_struct PMPI_Type_create_struct
-
-int MPIR_Type_create_struct_impl(int count,
-                                 const int array_of_blocklengths[],
-                                 const MPI_Aint array_of_displacements[],
-                                 const MPI_Datatype array_of_types[], MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int i, *ints;
-    MPI_Datatype new_handle;
-    MPIR_Datatype *new_dtp;
-    MPIR_CHKLMEM_DECL(1);
-
-    mpi_errno = MPIR_Type_struct(count,
-                                 array_of_blocklengths,
-                                 array_of_displacements, array_of_types, &new_handle);
-
-    MPIR_ERR_CHECK(mpi_errno);
-
-
-    MPIR_CHKLMEM_MALLOC_ORJUMP(ints, int *, (count + 1) * sizeof(int), mpi_errno,
-                               "content description", MPL_MEM_BUFFER);
-
-    ints[0] = count;
-    for (i = 0; i < count; i++)
-        ints[i + 1] = array_of_blocklengths[i];
-
-    MPIR_Datatype_get_ptr(new_handle, new_dtp);
-    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_STRUCT, count + 1,     /* ints (cnt,blklen) */
-                                           count,       /* aints (disps) */
-                                           count,       /* types */
-                                           ints, array_of_displacements, array_of_types);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
-
-  fn_exit:
-    MPIR_CHKLMEM_FREEALL();
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
 
 /*@
