+++ src/mpi/datatype/type_hvector.c
@@ -23,37 +23,6 @@ int MPI_Type_hvector(int count, int blocklength, MPI_Aint stride, MPI_Datatype o
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Type_hvector
 #define MPI_Type_hvector PMPI_Type_hvector
-
-int MPIR_Type_hvector_impl(int count, int blocklength, MPI_Aint stride, MPI_Datatype oldtype,
-                           MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPI_Datatype new_handle;
-    MPIR_Datatype *new_dtp;
-    int ints[2];
-
-    mpi_errno = MPIR_Type_vector(count, blocklength, (MPI_Aint) stride, 1,      /* stride in bytes */
-                                 oldtype, &new_handle);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    ints[0] = count;
-    ints[1] = blocklength;
-    MPIR_Datatype_get_ptr(new_handle, new_dtp);
-    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_HVECTOR, 2,    /* ints (count, blocklength) */
-                                           1,   /* aints */
-                                           1,   /* types */
-                                           ints, &stride, &oldtype);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-
 #endif
 
 /*@
